# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kang Raling API Web Service",
...
        "contact": {
            "name": "Developer Website Kang Raling",
            "email": "danu3anugrah@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server Lokal Development"
        },
        {
            "url": "https://kangraling.dlhgarut.id",
            "description": "Server Utama (Production Dinas Lingkungan Hidup)"
        }
    ],
    "paths": {
...
        "/artikels": {
            "get": {
                "tags": [
                    "Publik"
                ],
                "summary": "Ambil semua artikel yang dipublikasikan",
                "description": "Mengembalikan daftar artikel yang sudah dipublikasikan (`is_published: true`) untuk ditampilkan di halaman publik. Tidak membutuhkan token.",
                "operationId": "65d16f43aa0cc9da025baf4a3e4c9b5d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar artikel berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "judul": {
                                                        "type": "string",
                                                        "example": "Cara Memilah Sampah di Rumah"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "cara-memilah-sampah-di-rumah"
                                                    },
                                                    "gambar": {
                                                        "type": "string",
                                                        "example": "artikel/foto.jpg",
                                                        "nullable": true
                                                    },
                                                    "is_published": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 8
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Artikel"
                ],
                "summary": "Tambah artikel baru",
                "description": "Menambahkan artikel baru. Request menggunakan `multipart/form-data` karena mendukung upload gambar.",
                "operationId": "ff3951a75214621aa28a4fc2c861b79a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "isi_artikel"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "example": "Cara Memilah Sampah di Rumah"
                                    },
                                    "isi_artikel": {
                                        "type": "string",
                                        "example": "Isi artikel lengkap di sini..."
                                    },
                                    "gambar": {
                                        "description": "File gambar (maks 10MB)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artikel berhasil ditambahkan"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Tidak punya permission kelola.artikel"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/artikels/{slug}": {
            "get": {
                "tags": [
                    "Publik"
                ],
                "summary": "Ambil detail artikel berdasarkan slug",
                "description": "Mengembalikan detail lengkap satu artikel berdasarkan slug. Digunakan untuk halaman detail artikel publik.",
                "operationId": "98774d8812a66f2ca7b2694253973315",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug artikel",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "cara-memilah-sampah-di-rumah"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail artikel berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "judul": {
                                                    "type": "string",
                                                    "example": "Cara Memilah Sampah di Rumah"
                                                },
                                                "isi_artikel": {
                                                    "type": "string",
                                                    "example": "Isi artikel lengkap..."
                                                },
                                                "gambar": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "user": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artikel tidak ditemukan"
                    }
                }
            }
        },
        "/artikels/{id}/edit": {
            "get": {
                "tags": [
                    "Artikel"
                ],
                "summary": "Ambil data artikel untuk form edit (dashboard)",
                "description": "Mengambil data artikel berdasarkan ID untuk keperluan form edit di dashboard. Fasilitator hanya bisa akses artikel miliknya sendiri.",
                "operationId": "0010498d4aa4a9537fdf2bf41e56293c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID artikel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data artikel untuk edit berhasil diambil"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Fasilitator mencoba edit artikel milik orang lain"
                    },
                    "404": {
                        "description": "Artikel tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/artikels/{id}": {
            "post": {
                "tags": [
                    "Artikel"
                ],
                "summary": "Update artikel (gunakan method spoofing _method=PUT)",
                "description": "Mengupdate artikel yang sudah ada. Karena ada upload file, gunakan `multipart/form-data` dengan tambahan field `_method: PUT` untuk method spoofing.",
                "operationId": "7502ceec95ca84fad198cf8f54f6461d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "isi_artikel",
                                    "_method"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "Wajib diisi PUT untuk method spoofing",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "judul": {
                                        "type": "string",
                                        "example": "Cara Memilah Sampah di Rumah"
                                    },
                                    "isi_artikel": {
                                        "type": "string",
                                        "example": "Isi artikel yang diperbarui..."
                                    },
                                    "gambar": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artikel berhasil diperbarui"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    },
                    "404": {
                        "description": "Artikel tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Artikel"
                ],
                "summary": "Hapus artikel",
                "description": "Menghapus artikel secara permanen. Fasilitator hanya bisa hapus artikel miliknya sendiri.",
                "operationId": "0936de3b6be5d8d671f0ba87a0dba79e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artikel berhasil dihapus"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    },
                    "404": {
                        "description": "Artikel tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token dengan format: **Bearer {token}**",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}